принимает на вход число N и выдаёт произведение чисел от 1 до N

+++ task_recursion/Program.cs
@@ -2,9 +2,9 @@
 Console.Clear();
 Console.WriteLine("Введите число: ");
 int N = Convert.ToInt32(Console.ReadLine());
-int Product(int N)
+double Product(int N)
 {
     if (N == 1) return 1;
-    else return N * Product(N-1);
+    else return N * Product (N - 1);
 }
 Console.WriteLine(Product(N));
